fix: file mangement and homepage
